+++ Frontend/src/functions/auth.js
@@ -8,5 +8,14 @@ export const currentUser = async (userId) => {
       withCredentials: true,
     }
   );
-  
+};
+
+export const currentAdmin = async (userId) => {
+    return await axios.post(
+    `${import.meta.env.VITE_URL_API}/api/user/current-admin`,
+    {userId},
+    {
+      withCredentials: true,
+    }
+  );
 };
