+++ Frontend/src/pages/404/NotFound404.jsx
@@ -22,7 +22,7 @@ function Copyright() {
 // TODO remove, this demo shouldn't need to reset the theme.
 const defaultTheme = createTheme();
 
-export default function StickyFooter() {
+export default function NotFound404() {
   return (
     <ThemeProvider theme={defaultTheme}>
       <Box
