+++ Frontend/src/pages/admin/ManageUsers.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function ManageUsers() {
+  return (
+    <div>ManageUsers</div>
+  )
+}
