--- Frontend/src/pages/report/HeaderBar.jsx
@@ -1,68 +1,0 @@
-import react, { useState } from "react";
-import { Box, IconButton, Menu, MenuItem } from "@mui/material";
-import InputBase from "@mui/material/InputBase";
-import NotificationsOutlinedIcon from "@mui/icons-material/NotificationsOutlined";
-import SettingsOutlinedIcon from "@mui/icons-material/SettingsOutlined";
-import PersonOutlinedIcon from "@mui/icons-material/PersonOutlined";
-import SearchIcon from "@mui/icons-material/Search";
-import { Link } from "react-router-dom";
-
-const HeaderBar = () => {
-  const [anchorEl, setAnchorEl] = useState(null);
-
-  const handleMenu = (event) => {
-    setAnchorEl(event.currentTarget);
-  };
-
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
-  return (
-    <Box display="flex" justifyContent="space-between" p={2}>
-      {/* search  */}
-      <Box display="flex" borderRadius="3px" backgroundColor="#F5EFE7">
-        <InputBase sx={{ ml: 2, flex: 1 }} placeholder="Search" />
-        <IconButton type="button" sx={{ p: 1 }}>
-          <SearchIcon />
-        </IconButton>
-      </Box>
-
-      {/* icons */}
-      <Box display="flex">
-        <IconButton>
-          <NotificationsOutlinedIcon />
-        </IconButton>
-        <IconButton>
-          <SettingsOutlinedIcon />
-        </IconButton>
-        <IconButton  onClick={handleMenu}>
-          <PersonOutlinedIcon />
-          <Menu
-            id="menu-appbar"
-            anchorEl={anchorEl}
-            anchorOrigin={{
-              vertical: "top",
-              horizontal: "right",
-            }}
-            keepMounted
-            transformOrigin={{
-              vertical: "top",
-              horizontal: "right",
-            }}
-            open={Boolean(anchorEl)}
-            onClose={handleClose}
-          >
-            <Link to="#" className="menu-bars">
-              <MenuItem onClick={handleClose}>Profile</MenuItem>
-            </Link>
-            <Link to="#" className="menu-bars">
-              <MenuItem onClick={handleClose}>My account</MenuItem>
-            </Link>
-          </Menu>
-        </IconButton>
-      </Box>
-    </Box>
-  );
-};
-
-export default HeaderBar;
