--- Frontend/src/pages/report/SideBar.jsx
@@ -1,188 +1,0 @@
-import React, { useState } from "react";
-import { Sidebar, Menu, MenuItem, SubMenu } from "react-pro-sidebar";
-import { Box, IconButton, Typography, useTheme, Badge } from "@mui/material";
-import { Link } from "react-router-dom";
-import HomeOutlinedIcon from "@mui/icons-material/HomeOutlined";
-import PeopleOutlinedIcon from "@mui/icons-material/PeopleOutlined";
-import ReceiptOutlinedIcon from "@mui/icons-material/ReceiptOutlined";
-import CalendarTodayOutlinedIcon from "@mui/icons-material/CalendarTodayOutlined";
-import BarChartOutlinedIcon from "@mui/icons-material/BarChartOutlined";
-import MenuOutlinedIcon from "@mui/icons-material/MenuOutlined";
-import MapOutlinedIcon from "@mui/icons-material/MapOutlined";
-import TableViewIcon from "@mui/icons-material/TableView";
-import Person from "@mui/icons-material/Person";
-
-import Users from "./user/Users";
-import Questionnaires from "./questionnaires/Questionnaires";
-import ContentMap from "./ContentMap";
-
-const SideBar = ({ isSidebar, updateSidebarData }) => {
-  const [isCollapsed, setisCollapsed] = useState(false);
-  const [toggled, setToggled] = useState(false);
-  const [broken, setBroken] = useState(false);
-  const [selectedMenuItem, setSelectedMenuItem] = React.useState(null);
-
-  // const handleClick = () => {
-  //   // Example: send data back to MainApp component
-  //   const newData = "New Data from SideBar";
-  //   updateSidebarData(newData);
-  // };
-
-  const handleMenuItemClick = (menuItem) => {
-    setSelectedMenuItem(menuItem);
-    /// props
-    updateSidebarData(menuItem)
-
-  };
-
-  const renderComponent = () => {
-    console.log('sdas', selectedMenuItem)
-
-  };
-
-  return (
-    <div
-      style={{
-        display: "flex",
-        height: "100%",
-      }}
-    >
-      <Sidebar
-        collapsed={isCollapsed}
-        toggled={toggled}
-        onBackdropClick={() => setToggled(false)}
-        onBreakPoint={setBroken}
-        image="/public/logo-nano.png"
-        breakPoint="md"
-        style={{ height: "100%" }}
-      >
-        <div
-          style={{ display: "flex", flexDirection: "column", height: "100%" }}
-        >
-          <div style={{ flex: 1, marginBottom: "32px" }}>
-            <Menu iconShape="square">
-              {/* LOGO */}
-              <MenuItem
-                onClick={() => setisCollapsed(!isCollapsed)}
-                icon={isCollapsed ? <MenuOutlinedIcon /> : undefined}
-                style={{
-                  margin: "10px 0 20px 0",
-                }}
-              >
-                {!isCollapsed && (
-                  <Box
-                    display="flex"
-                    justifyContent="space-between"
-                    alignItems="center"
-                    ml="15px"
-                  >
-                    <Typography>NANOET</Typography>
-                    <IconButton onClick={() => setisCollapsed(!isCollapsed)}>
-                      <MenuOutlinedIcon />
-                    </IconButton>
-                  </Box>
-                )}
-              </MenuItem>
-              {!isCollapsed && (
-                <Box mb="25px">
-                  <Box
-                    display="flex"
-                    justifyContent="center"
-                    alignItems="center"
-                  >
-                    <img
-                      alt="profile-user"
-                      width="100px"
-                      height="100px"
-                      src={`/assets/user.jpg`}
-                      style={{ cursor: "pointer", borderRadius: "50%" }}
-                    />
-                  </Box>
-                  <Box textAlign="center">
-                    <Typography sx={{ m: "10px 0 0 0" }}>REPORT</Typography>
-
-                  </Box>
-                </Box>
-              )}
-
-              {/* <Link to="/admin" className="menu-bars"> */}
-              <MenuItem
-                icon={<HomeOutlinedIcon />}
-
-              >
-                Dashboard
-              </MenuItem>
-              {/* </Link> */}
-
-              <SubMenu icon={<MapOutlinedIcon />} label="ข้อมูล">
-
-                <Link to={"/report/question"} className="menu-bars">
-                  <MenuItem icon={<TableViewIcon />} > คำถาม</MenuItem>
-                </Link>
-
-
-                <Link to={"/report/questionnaires"} className="menu-bars">
-                  <MenuItem icon={<BarChartOutlinedIcon />}>
-                    {" "}
-                    แบบสอบถาม
-                  </MenuItem>
-                </Link>
-
-                <Link to={"/report/users"} className="menu-bars">
-                <MenuItem icon={<Person />}>
-                  {" "}
-                  ผู้ใช้งาน
-                </MenuItem>
-                </Link>
-              </SubMenu>
-
-              <SubMenu label="Manage" icon={<PeopleOutlinedIcon />}>
-                <Link to={"/admin/manage"} className="menu-bars">
-                  <MenuItem>User</MenuItem>
-                </Link>
-                <MenuItem> Admin</MenuItem>
-              </SubMenu>
-            </Menu>
-
-            <div
-              style={{
-                padding: "0 24px",
-                marginBottom: "8px",
-                marginTop: "32px",
-              }}
-            >
-              <Typography
-                variant="body2"
-                fontWeight={600}
-                style={{
-                  opacity: isCollapsed ? 0 : 0.5,
-                  letterSpacing: "0.5px",
-                }}
-              >
-                Extra
-              </Typography>
-            </div>
-
-            <Menu>
-              <MenuItem icon={<CalendarTodayOutlinedIcon />}>Calendar</MenuItem>
-              <MenuItem icon={<ReceiptOutlinedIcon />}>Documentation</MenuItem>
-            </Menu>
-          </div>
-        </div>
-      </Sidebar>
-      <main>
-        <div style={{ padding: "16px 2px ", color: "#44596e" }}>
-          <div style={{ marginBottom: "16px" }}>
-            {broken && (
-              <IconButton onClick={() => setToggled(!toggled)}>
-                <MenuOutlinedIcon />
-              </IconButton>
-            )}
-          </div>
-        </div>
-        {renderComponent()}
-      </main>
-    </div>
-  );
-};
-export default SideBar;
