--- Frontend/src/pages/report/question/Question.css
@@ -1,13 +1,0 @@
-.css-1ygcj2i-MuiTableCell-root {
-    font-family: "Roboto","Helvetica","Arial",sans-serif;
-    font-weight: 500;
-    font-size: 0.875rem;
-    line-height: 1.5rem;
-    letter-spacing: 0.01071em;
-    display: table-cell;
-    vertical-align: inherit;
-    border-bottom: 1px solid rgba(224, 224, 224, 1);
-    text-align: center;
-    padding: 16px;
-    color: rgba(0, 0, 0, 0.87);
-}
