--- Frontend/src/pages/report/user/Table.css
@@ -1,84 +1,0 @@
-body {
- 
-  align-items: center;
-  justify-content: center;
-  height: 100vh;
-  margin: 0;
-}
-
-.image-box {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-wrap: wrap; /* กรณีที่มีการ์ดหลายๆ แถว */
-}
-
-.card-box{
- 
-  width: 25rem;
-  justify-content: center;
-  align-items: center;
-  margin: 40px;
-  margin-bottom: 10px;
-}
-
-.text-if {
-  justify-content: center;
-}
-
-.button-container {
-  display: flex;
-  justify-content: space-around;
-  margin-top: 20px;
-}
-
-.save-button {
-  padding: 20px 40px;
-  font-size: 26px;
-  background-color: #0b9539;
-  border: 1px solid #ccc;
-  color: rgb(13, 13, 13);
-  border-radius: 35px;
-  box-shadow: 4px 8px 3px 5px #b2b2b2;
-  font-size: larger;
-  cursor: pointer;
-  text-align :center;
-  margin: 25px;
-  font-weight: bold;
-}
-
-.save-button:active{
-  background-color: #0b9539;
-  color: rgb(255, 255, 255);
-  transform: translateY(10px);
-  outline: none;
-}
-
-.edit-button {
-  padding: 20px 40px;
-  font-size: 26px;
-  background-color: #dfae4d;
-  border: 1px solid #ccc;
-  color: rgb(13, 13, 13);
-  border-radius: 35px;
-  box-shadow: 4px 8px 3px 5px #b2b2b2;
-  font-size: larger;
-  cursor: pointer;
-  text-align :center;
-  margin: 25px;
-  font-weight: bold;
-}
-
-.edit-button:active{
-  background-color: #dfae4d;
-  color: rgb(255, 255, 255);
-  transform: translateY(10px);
-  outline: none;
-}
-
-@media screen and (max-width: 400px) {
-  .button-container {
-    flex-direction: column; /* จัดวางปุ่มในคอลัมน์เมื่อหน้าจอมีขนาดน้อยกว่า 400px */
-    align-items: center; /* จัดให้ปุ่มอยู่ตรงกลาง */
-  }
-}
